refactor(remote): refactor StreamContext into lib functions
to allow benches...
